Add files via upload
In this project I created an easy-to-use and completely responsive Online Meal Prep Service Website for people to order healthy and balanced meal plans according to their dietary needs. The central goal of this project is to maximize efficiency and optimize convenience to provide a hassle-free experience to the target group. We present a website with multiple meal plans helping people to choose what they desire. The website has been built using HTML, CSS, JavaScript, and PHP. Machine learning has been incorporated into the website with the use of Python. We have included multiple features to help users analyze and construct their meal plans in three different ways. In the first meal plan option, the user is free to choose whatever meal options they require. In the second meal plan option, the user is provided fixed meal options according to Body Mass Index (BMI) and the Mifflin-St Jeor Equation. In the third meal plan, users are recommended meal options with the help of machine learning. Our objective is to help the user have a healthy and balanced meal plan and a convenient experience while using the website.